requirements.txt
Django==4.1.2
django-widget-tweaks==1.4.12
Faker==15.1.1
phonenumbers==8.13.5
django-phonenumbers==1.0.1
django-phonenumber-field==7.0.2
django-countries==7.5.1

